Add listcell viewmodel
